+++ CryptocurrencyExchange/CryptocurrencyExchange/Sources/DataSources/Network/WebSocket/Bithumb/ResponseDTO/BithumbTransactionSocketResponseDTO.swift
@@ -16,11 +16,11 @@ struct BithumbTransactionSocketResponseDTO: Decodable {
     struct Transactions: Codable {
         let list: [Transaction]
     }
-
+    
     struct Transaction: Codable {
         let symbol, buySellGB, contPrice, contQty: String
         let contAmt, contDtm, updn: String
-
+        
         enum CodingKeys: String, CodingKey {
             case symbol
             case buySellGB = "buySellGb"
@@ -40,15 +40,15 @@ extension BithumbTransactionSocketResponseDTO {
 extension BithumbTransactionSocketResponseDTO.Transaction {
     func toDomain() -> BithumbTransactionStream {
         return BithumbTransactionStream(
-           symbol: symbol,
-           transactionType: BithumbTransactionType(
-            rawValue: Int(buySellGB) ?? -1
-           ) ?? .none,
-           contPrice: Double(contPrice) ?? 0,
-           contQuantity: Double(contQty) ?? 0,
-           contAmount: Double(contAmt) ?? 0,
-           transactionDate: contDtm.toDate(format: "yyyy-MM-dd HH:mm:ss.SSSSSS") ?? Date(),
-           priceUpDown: BithumbTransactionPriceUpDown(rawValue: updn) ?? .none
+            symbol: symbol,
+            transactionType: BithumbTransactionType.from(
+                value: Int(buySellGB) ?? -1
+            ),
+            contPrice: Double(contPrice) ?? 0,
+            contQuantity: Double(contQty) ?? 0,
+            contAmount: Double(contAmt) ?? 0,
+            transactionDate: contDtm.toDate(format: "yyyy-MM-dd HH:mm:ss.SSSSSS") ?? Date(),
+            priceUpDown: BithumbTransactionPriceUpDown(rawValue: updn) ?? .none
         )
     }
 }
