+++ CryptocurrencyExchange/CryptocurrencyExchange/Sources/Presentations/Scene/Topkim/TransactionListEnviroment.swift
@@ -1,0 +1,13 @@
+//
+//  TransactionListEnviroment.swift
+//  CryptocurrencyExchange
+//
+//  Created by 김정상 on 2022/03/11.
+//
+
+import Foundation
+
+struct TransactionListEnvironment {
+    let transactionListUseCase: () -> TransactionListUseCaseProtocol
+    let toastClient: ToastClient
+}
